third_party/blink/web_tests/external/wpt/html/rendering/non-replaced-elements/the-fieldset-and-legend-elements/legend-align-justify-self-expected.txt

This is a testharness.js-based test.
[FAIL] <fieldset><legend align="left">x</legend></fieldset>
  assert_equals: expected "left" but got "auto"
[FAIL] <fieldset><legend align="center">x</legend></fieldset>
  assert_equals: expected "center" but got "auto"
[FAIL] <fieldset><legend align="right">x</legend></fieldset>
  assert_equals: expected "right" but got "auto"
[FAIL] <fieldset><legend align="lEfT">x</legend></fieldset>
  assert_equals: expected "left" but got "auto"
[FAIL] <fieldset><legend align="cEnTeR">x</legend></fieldset>
  assert_equals: expected "center" but got "auto"
[FAIL] <fieldset><legend align="rIgHt">x</legend></fieldset>
  assert_equals: expected "right" but got "auto"
Harness: the test ran to completion.
